+++ lib/support-relationships/index.ts
@@ -1,1 +1,3 @@
 export * from "./supportRelationships";
+
+export * from "./relationshipActivation";